Apply validations on events if defined.
As Events are posted, we can apply validations if a "validator" exists.
A "Validate#{event.category}".constantize naming convention is required.

For example, an event is POSTed with a category of `til_tweet`.  We
don't have any validators for that, so it gets created (assuming it
passes AR validations).  On the other hand, if an event with
`resume_site` gets POSTed and we have a validator class called
ValidateResumeSite, it will be executed.

All validators should have the `Interactor` module included.
The intention is define all future validators in a `interactions/validations`
directory.